# docs/swagger.yaml
basePath: /api
definitions:
  handlers.BulkImportRequest:
    properties:
      chapters:
        items:
          properties:
            content:
              type: string
            title:
              type: string
          required:
          - content
          - title
          type: object
        minItems: 1
        type: array
    required:
    - chapters
    type: object
...
  handlers.CreateChapterRequest:
    properties:
      content:
        type: string
      title:
        type: string
    required:
    - content
    - title
    type: object
  handlers.CreateCommentRequest:
    properties:
      chapter_id:
        description: Optional, nếu comment cho chapter cụ thể
        type: string
      content:
        maxLength: 2000
        type: string
    required:
    - content
    type: object
...
  handlers.ReplyCommentRequest:
    properties:
      content:
        maxLength: 2000
        type: string
    required:
    - content
    type: object
  handlers.ScheduleChapterRequest:
    properties:
      scheduled_at:
        description: RFC3339 format
        type: string
    required:
    - scheduled_at
    type: object
  response.Meta:
    properties:
      limit:
        type: integer
      page:
        type: integer
      total:
        type: integer
      total_pages:
        type: integer
    type: object
  response.Pagination:
    properties:
      data: {}
      meta:
        $ref: '#/definitions/response.Meta'
      success:
        type: boolean
    type: object
  response.Response:
    properties:
      data: {}
      error:
        type: string
      message:
        type: string
      success:
        type: boolean
    type: object
host: localhost:8080
info:
  contact: {}
  description: API cho nền tảng đọc truyện web novel Nekozanedex
  title: Nekozanedex API
  version: "1.0"
paths:
  /api/admin/chapters/{id}:
    delete:
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Xóa chapter (Admin)
      tags:
      - Admin - Chapters
    get:
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Lấy chapter theo ID (Admin)
      tags:
      - Admin - Chapters
    put:
      consumes:
      - application/json
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      - description: Chapter Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChapterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Cập nhật chapter (Admin)
      tags:
      - Admin - Chapters
  /api/admin/chapters/{id}/publish:
    post:
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Xuất bản chapter (Admin)
      tags:
      - Admin - Chapters
  /api/admin/chapters/{id}/schedule:
    post:
      consumes:
      - application/json
      parameters:
      - description: Chapter ID
        in: path
        name: id
        required: true
        type: string
      - description: Schedule Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ScheduleChapterRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Hẹn giờ xuất bản chapter (Admin)
      tags:
      - Admin - Chapters
...
  /api/admin/stories/{storyId}/chapters:
    post:
      consumes:
      - application/json
      parameters:
      - description: Story ID
        in: path
        name: storyId
        required: true
        type: string
      - description: Chapter Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateChapterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tạo chapter mới (Admin)
      tags:
      - Admin - Chapters
  /api/admin/stories/{storyId}/chapters/bulk:
    post:
      consumes:
      - application/json
      parameters:
      - description: Story ID
        in: path
        name: storyId
        required: true
        type: string
      - description: Chapters Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.BulkImportRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Import nhiều chapters cùng lúc (Admin)
      tags:
      - Admin - Chapters
...
  /api/chapters/{chapterId}/comments:
    get:
      parameters:
      - description: Chapter ID
        in: path
        name: chapterId
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      summary: Lấy comments của chapter
      tags:
      - Comments
  /api/comments/{commentId}:
    delete:
      parameters:
      - description: Comment ID
        in: path
        name: commentId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Xóa comment
      tags:
      - Comments
  /api/comments/{commentId}/reply:
    post:
      consumes:
      - application/json
      parameters:
      - description: Parent Comment ID
        in: path
        name: commentId
        required: true
        type: string
      - description: Reply Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.ReplyCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Trả lời comment
      tags:
      - Comments
...
  /api/stories/{slug}/chapters:
    get:
      parameters:
      - description: Story Slug
        in: path
        name: slug
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Lấy danh sách chapters của truyện
      tags:
      - Chapters
  /api/stories/{slug}/chapters/{number}:
    get:
      parameters:
      - description: Story Slug
        in: path
        name: slug
        required: true
        type: string
      - description: Chapter Number
        in: path
        name: number
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: Lấy chapter theo số
      tags:
      - Chapters
  /api/stories/{slug}/comments:
    get:
      parameters:
      - description: Story Slug
        in: path
        name: slug
        required: true
        type: string
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 20
        description: Items per page
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Pagination'
      summary: Lấy comments của truyện
      tags:
      - Comments
  /api/stories/{storyId}/comments:
    post:
      consumes:
      - application/json
      parameters:
      - description: Story ID
        in: path
        name: storyId
        required: true
        type: string
      - description: Comment Info
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/handlers.CreateCommentRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - BearerAuth: []
      summary: Tạo comment
      tags:
      - Comments
...
securityDefinitions:
  BearerAuth:
    description: 'Bearer token authentication. Format: "Bearer {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"